0-intro/pyproject.toml
[project]
name = "0-intro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "smolagents>=1.3.0",
]
